Added preliminary support for donations!
AFAIK, the support should be there, just need to add the items once Callisto is in the market.

There was SOME support added for IRC, but I'm thinking of switching the API I'm using to PircBot,
so it will be put on hold with jerklib and the branch for Pirc will begin.