Fixed streaming!!! (for the most part)
Apparently my LG Optimus Slider is just retarded because it works on my older LG Optimus Prime.
However, in the event that it doesn't work, it will send me an anonymous error report
with the verion of Callisto and the version of Android the user is running.

Also, made Callisto compatible with down to Android 1.5! (Mostly just switching 'match_parent' to 'fill_parent')

This was supposed to be the 0.7 Beta release, but I got sick a few days ago and so will miss my deadline. Oh well.