This branch is for trying to remove the LiveStream page.
Basically, it's not needed at all. Taking out the image will shrink the apk.
However, I don't want it to be JUST the button on the home screen. Then, if you accidentally press the live button, it will
start playing.

I also removed a lot of the donation bonuses in an attempt to shrink the APK. I should have done that on the main branch,
but it didn't occur to me until later.

I also removed all of the Android Billing files (the code is still in there to make it work, the files just now won't be
taking up space in the APK).