Switched to Jerklib extracted.
The plan is to fix many of the current issues with how IRC is treated in Callisto.
    -Timeout: For when a connection can be made but is taking forever
    -Retries: For when no connection can be made
    -Posting messages to the wrong screen: I am using two different Spanned queues instead of one shared Spanned object.
     Hopefully this means that they will always be posted in the right order and on the right screen.

The only other thing I can think of to fix would be the NickServ problem, which is next.

NOTE: Hopefully this is temporary. But I found some things in JerkLib either lacking or not well enough explained in the
docs.
For example, how it does not attempt to check that a connection to the server exists befor issuing an IRC command.