VASTLY improved the contact form.
Instead of doing all the HTTP work and then looping over it in a buffer and shit, now it:
	* Requests the original form
	* Signals a Javascript handler once the page has finished loading
	* Gets the page source inside the Javascript handler
	* Rips out the CSS and replaces it with the custom
	* Loads the source with a base URL.

This allows it to work perfectly with the "error" page, as well as shrinking the code size and making it 100x easier to read.
Not too much of a performance boost at this point, seems about the same.

Also, to avoid showing the un-formatted page beforehand right now it hides the WebView until it has loaded the acquired source.
This works ok, but there is also a brief flicker for some reason that I can't get rid of.

Oh, and I also formatted the error page as well. It looks really nice, even highlights the parts of the form you left blank.

So proud of it right now.