Re-wrote the way a show is re-freshed.
WAY better. Now it searches within the scope of the <item> tags, and does not expect the elements to be in any certain order.

This was to solve a problem with the new "Plan B" feed, in that NOTHING WAS IN THE RIGHT ORDER.

Anyway this is more reliable. Cleaner, easier to read, and just more compartmentalized.