TRYING unsuccessfully to fix a big in Android
There's a bug in Android 2.x that a PreferenceActivity can
be skinned, but not any PreferenceScreen's that it contains.
So the result is you get black text on black bg.

There are workarounds for this BUT the problem is when you rotate
your device, you have no way of accessing the Dialog created by the
preferencescreen.